sched: Add unweighted load based on running busy time to PELT
With the changed PELT framework in place, per-entity load tracking for
unweighted load based on running busy time is a simple and low-overhead
addition alongside the already existing implementation for weighted load
based on runnable busy time.

Since for load based on running busy time tracking we have to be able to
distinguish if a task is running or not (waiting or blocking). To be able
to do so, another call to update_entity_load_avg has been introduced into
set_next_entity.

Related to __update_group_entity_contrib, unlike the contribution of
weighted load based on runnable busy time, the contribution of unweighted
load based on running time for a group entity per cpu's sched entity
{tg->se[cpu]} is the sum of the contribution of unweighted load based on
running time for all sched entities on the group's cfs run queue
{tg->cfs_rq[cpu]}. It is _not_ influenced in any way by the task group
h_load.

Hence it is representing the actual cpu usage of the group, not
its intended load contribution which may differ significantly from the
usage on lightly utilized systems.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>