sched: sched feature for cpu frequency selection
This patch introduces the SCHED_ENERGY_FREQ sched feature, which is
implemented using jump labels when SCHED_DEBUG is defined. It is
statically set to false when SCHED_DEBUG is not defined and thus
disabled by default.

Signed-off-by: Michael Turquette <mturquette@linaro.org>

Conflicts:
	kernel/sched/fair.c
	kernel/sched/features.h
Trivial fixes required by the fact that EAS code introduces an
additional sched_feature and cpu_overutilized() function.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>