blmp: domain and task selection utilities
Introduces all the support methods to select tasks, to be moved among
domains, and target domains, where selected tasks should be move to.

These selection methods implement some heuristics which allows to:

* select the heaviest/lightest task in the first 5 runnable of a runqueue
* get the load ratio of the lowest-loaded CPU in an hmp_domain
* select a slower/faster CPU suitable to move a task into

A next up/down migration delay is used to prevent tasks from doing another
migration in the same direction until a delay has expired and eventually
tasks load adapted to the compute capacity of the new domain.

In case of multiple CPUs with the same load ratio, the one which is the
last-recently-disturbed (LRD) is selected.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Liviu Dudau <liviu.dudau@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>