blmp: migration strategy: WAKEUP
Goals:
1. keep just a single task on each CPU of the faster domain

Once a task wakeups this patch forces its migration to a faster CPU, if
it meets the requirements imporsed by "hmp_up_migration", of a slower
CPU, if it meets the requirements imposed by the "hmp_down_migration".

In any other case the task is not allowed to migrate to a CPU which is
not part of its current domain.

Signed-off-by: Chris Redpath <chris.redpath@arm.com>
Signed-off-by: Robin Randhawa <robin.randhawa@arm.com>
Signed-off-by: Liviu Dudau <Liviu.Dudau@arm.com>
Signed-off-by: Jon Medhurst <tixy@linaro.org>