sched: Make energy awareness a sched feature
Was: [RFCv2 PATCH 02/23] sched: Make energy awareness a sched feature

This patch introduces the ENERGY_AWARE sched feature, which is
implemented using jump labels when SCHED_DEBUG is defined. It is
statically set false when SCHED_DEBUG is not defined. Hence this doesn't
allow energy awareness to be enabled without SCHED_DEBUG. This
sched_feature knob will be replaced later with a more appropriate
control knob when things have matured a bit.

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>
[juril; backported to 3.10]
The ENERGY_AWARE sched feature is available for FAI_GROUP_SCHED
configs only, as EAS depends on PELT.
Signed-off-by: Juri Lelli <juri.lelli@arm.com>