sched: Energy model functions
Introduces energy_diff_util() which finds the energy impacts of adding
or removing utilization from a specific cpu. The calculation is based on
the energy information provided by the platform through sched_energy
data in the sched_domain hierarchy.

Task and cpu utilization is based on unweighted load tracking
(uw_load_avg_contrib) and unweighted cpu_load(cpu, 1) that is introduced
earlier in the patch set. While it isn't a perfect utilization metric,
it is better than weighted load. There are several other loose ends that
need to be addressed, such as load/utilization invariance and proper
representation of compute capacity. However, the energy model and
unweighted load metrics are there.

The energy cost model only considers utilization (busy time) and idle
energy (remaining time) for now. The basic idea is to determine the
energy cost at each level in the sched_domain hierarchy.

        for_each_domain(cpu, sd) {
                sg = sched_group_of(cpu)
                energy_before = curr_util(sg) * busy_power(sg)
                                + (1-curr_util(sg)) * idle_power(sg)
                energy_after = new_util(sg) * busy_power(sg)
                                + (1-new_util(sg)) * idle_power(sg)
                energy_diff += energy_before - energy_after
        }

Wake-ups energy is added later in this series.

Assumptions and the basic algorithm are described in the code comments.

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>

[juril; backported to 3.10]
More assumption are introduced in the backport. We are now based on
Vincent's v9, so we have a utilization signal (get_cpu_usage() and
se.avg.utilization_avg_contrib) that we can use instead of the old
uw_load_* signal (and variants, e.g. cpu_load()).
The fact that we don't have an uw_cpu_load array anymore introduces
assumption on the next coming patches (but, we should be fine with
it).
Moreover, all this won't work on single cluster systems, as we don't
backport the hack providing system-wide energy data.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>