sched: Use energy to guide wakeup task placement
Attempt to pick most energy efficient wakeup in find_idlest_{group,
cpu}(). Finding the optimum target requires an exhaustive search
through all cpus in the groups. Instead, the target group is determined
based on load and probing the energy cost on a single cpu in each group.
The target cpu is the cpu with the lowest energy cost.

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>

[juril; backported to 3.10]
As noted in "sched: Energy model functions" we here have the big
assumption that we never use load_idx != 0 for source{target}_load in
find_target_group. A WARN_ON() is added to check if we break this
assumption.
Signed-off-by: Juri Lelli <juri.lelli@arm.com>
[mortenr; Rebased on tip]