cpuidle: Store the idle start time stamp
The scheduler uses the idle timestamp stored in the struct rq to retrieve the
time when the cpu went idle in order to find the idlest cpu. Unfortunately
this information is wrong as it does not have the same meaning from the cpuidle
point of view. The idle_stamp in the struct rq gives the information when the
idle task has been scheduled while the idle task could be interrupted several
times and the cpu going through an idle/wakeup multiple times.

Add the idle start time in the idle state structure.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>