cpuidle: menu: Remove the default poll state
The following applies only for x86.

The current code does default to C0, only if a timer is about to expire.

As the timer expiration is a reliable information, this has a double guarantee:
 * we ensure fast exit from idle
 * we ensure we won't be polling for a too long period which is dangerous from
   a thermal point of view

Unfortunately this code brings a lot of weirdness all around the default idle
state and with the CPUIDLE_DRIVER_STATE_START macro.

Regarding the number of times the poll function is called (1/10000 on my
server), we can legitimately ask if this test is worth in the menu governor
with the hardware we have nowadays with very low exit latency.

Just remove this test. If it appears it brings a real measurable gain in
performance, we can re-introduce it in a more clever way later.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>