cpuidle: Remove CPUIDLE_DRIVER_STATE_START users from cpuidle.c
In order to remove the poll state from the idle drivers in the next patch,
let's remove the usage of the CPUIDLE_DRIVER_STATE_START from the idle state
selection loops.

1. cpuidle_play_dead will ignore the poll idle state because this one does not
implement the 'enter_dead' callback. We can safely remove the usage of the
CPUIDLE_DRIVER_STATE_START macro and start from the zero index.

2. cpuidle_find_deepest_state will always ignore the poll idle state because
its exit_latency is 0 and because of the check in the loop:

	if (s->disabled || su->disable || s->exit_latency <= latency_req)
		continue;

... it will be always ignored (exit_latency == latency_req)

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>