cpuidle: x86: Remove the poll state for the x86 driver
All the code is avoiding to use the poll state. The only place where the poll
state is used is when we failed to find an idle state and there is a timer to
expire within 5us. But in this place we call directly cpuidle_poll without
using the poll state callback in the idle state array of the cpuidle driver.

The poll state in the driveri's idle state array is no longer used.

Remove it and cleanup this mess.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>