arm: Convert kvm to hotplug state machine
Install the callbacks via the state machine and let the core invoke
the callbacks on the already online CPUs.
The error path in init_hyp_mode() seems broken. cpuhp_remove_state() is
added after the out_free_mappings label so it is always executed in case of
failure and it does not matter if it is invoked without a matching
cpuhp_setup_state() invocation - nothing will be done.

Cc: Gleb Natapov <gleb@kernel.org>
Cc: Paolo Bonzini <pbonzini@redhat.com>
Cc: Christoffer Dall <christoffer.dall@linaro.org>
Cc: Marc Zyngier <marc.zyngier@arm.com>
Cc: Russell King <linux@arm.linux.org.uk>
Cc: kvm@vger.kernel.org
Cc: kvmarm@lists.cs.columbia.edu
Reviewed-by: Sebastian Andrzej Siewior <bigeasy@linutronix.de>
Signed-off-by: Richard Cochran <rcochran@linutronix.de>