FROMLIST: sched: Calculate energy consumption of sched_group
For energy-aware load-balancing decisions it is necessary to know the
energy consumption estimates of groups of cpus. This patch introduces a
basic function, sched_group_energy(), which estimates the energy
consumption of the cpus in the group and any resources shared by the
members of the group.

NOTE: The function has five levels of identation and breaks the 80
character limit. Refactoring is necessary.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>

(am from https://patchwork.kernel.org/patch/6390611)
Signed-off-by: Juri Lelli <juri.lelli@arm.com>

BUG=chromium:48849
TEST=build and boot tested

Change-Id: If6015df53c26a3e86477f7b145fe87be889e9b23