irq_timings: add per-CPU prediction queueing
Once a good IRQ prediction is made, we need to enqueue it for later
consumption.  While at it we discard any predictions whose time stamp
is in the past.

There shouldn't be that many expected IRQs at any given time. A sorted
list is most likely going to be good enough.  And, by definition, the most
frequent IRQs will end up near the beginning of the list anyway.

Tthere is no generic way to determine what the IRQ controller is going
to do if the IRQ affinity mask contains multiple CPUs. It is therefore
assumed that the next occurrence of an IRQ is most likely to happen on
the same CPU as the last one. It appears to be the case overall from
observations on X86 despite active migration controlled from user space.
On ARM it is the first CPU in the affinity mask that is selected by the
GIC driver so this assumption is quite right in that case. If migration
frequency becomes significant compared to IRQ occurrences then we could
consider registering an affinity notifier.

Signed-off-by: Nicolas Pitre <nico@linaro.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>