irq_timings: function to retrieve time of next predicted IRQ
Those events in the past, if any, are purged and then the first item on
the list, if any, contains our next predicted IRQ time.

We have access to the standard deviation and could use it to qualify our
confidence in the prediction eventually. For now it's only the raw
prediction that is returned.

Signed-off-by: Nicolas Pitre <nico@linaro.org>
Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>