irq_timing: proc: Register irq to be tracked for timing measurements
The irq timing framework will track only the interrupts which were explicitely
specified with the flag IRQF_TIMING. That allows to have a fine grained control
of what we are tracking as source of events. Unfortunately each driver must be
modified accordingly and that makes very difficult the development.

This patch adds /proc/irq/<nr>/timing boolean file to specify from userspace the
interrupts we want to track. So no need to find and to change for each platform
the drivers in the kernel. That can be done from userspace now.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>