sched: Revert cpu scale-invariant load tracking
The review of RFCv4 on LKML has revealed that cpu scale-invariant
load tracking needs reworking of the load balance code to assure
that we still use normalized load and not cpu-scaled load directly.

This patch gets rid of cpu scale-invariance for load. So load is only
frequency scaled whereas utilization remains frequency and cpu scaled.

THIS ISSUE HAS TO BE RESOLVED IN AN RFCV4 REWORK.

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>