FROMLIST: sched: Determine the current sched_group idle-state
To estimate the energy consumption of a sched_group in
sched_group_energy() it is necessary to know which idle-state the group
is in when it is idle. For now, it is assumed that this is the current
idle-state (though it might be wrong). Based on the individual cpu
idle-states group_idle_state() finds the group idle-state.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>
Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>

(am from https://patchwork.kernel.org/patch/6390771)
Signed-off-by: Juri Lelli <juri.lelli@arm.com>

BUG=chromium:48849
TEST=build and boot tested

Change-Id: I2f895d5f9ed5371e1bcb101a51e1d70a5067fabe