sched: Introduce energy awareness into detach_tasks
Energy-aware load balancing does not rely on env->imbalance but instead it
evaluates the system-wide energy difference for each task on the src rq by
potentially moving it to the dst rq. If this energy difference is less
than zero the task is actually moved from src to dst rq.

Since the tasks are only removed from the src rq and not added to the dst
rq in the while loop of detach_tasks, usage_detached holds the sum of
usage from tasks already detached so by adding usage_detached and the
utilization of the task which is currently under investigation we get the
correct results from the energy_diff and __cpu_overutilized function calls.

cc: Ingo Molnar <mingo@redhat.com>
cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>