sched/fair: Make usage tracking cpu scale-invariant
Besides the existing frequency scale-invariance correction factor, apply
cpu scale-invariance correction factor to usage tracking to compensate
for any differences in compute capacity. This could be due to
micro-architectural differences (i.e. instructions per seconds) between
cpus in HMP systems (e.g. big.LITTLE), and/or differences in the maximum
frequency supported by individual cpus in SMP systems. Currently, usage
isn't comparable between cpus as it is relative to the capacity of each
individual cpu.

Each segment of the sched_avg::running_avg_sum geometric series is now
scaled by the cpu performance factor too so the
sched_avg::utilization_avg_contrib of each entity will be invariant from
the particular cpu of the HMP/SMP system on which the sched_entity is
scheduled.

With this patch, the usage level that is returned by get_cpu_usage()
stays relative to the max cpu performance of the fastest cpu in the
system.

In contrast to usage, load (sched_avg::runnable_avg_sum) should not be
scaled by compute capacity. The usage metric is based on running time
which only makes sense when cpus are _not_ fully utilized (usage cannot
go beyond 100% even if more tasks are added), where load is runnable
time which isn't limited by the capacity of the cpu and therefore is a
better metric for overloaded scenarios. If we run two busy loops on two
cpus with different compute capacity their load should be similar since
their compute demands are the same. We have to assume that the compute
demand of any task runs on a fully utilized cpu (no spare cycles = 100%
usage) is high and the same no matter the compute capacity of its
current cpu, hence we shouldn't scale load by cpu capacity.

Cc: Ingo Molnar <mingo@redhat.com>
Cc: Peter Zijlstra <peterz@infradead.org>

Signed-off-by: Dietmar Eggemann <dietmar.eggemann@arm.com>
Signed-off-by: Morten Rasmussen <morten.rasmussen@arm.com>