WIP: Sound: core driver
[Problem]
Currently hw_ptr will be increased by a buffer size if
1. expected hw ptr is larger then HW reported, and
2. time from last recorded time is over half a buffer time.
However, when HW reported pointer is only a few bytes smaller than
expected hw ptr, and period size is half of the buffer size,
this condition will be wrongly entered.

[Solution]
The original condition #2 should be wrong since in the end,
a buffer size is added to the hw ptr, not half a buffer size.
Only when the time difference is really over a buffer time then
one should add a buffer size to the hw ptr.

Signed-off-by: Koro Chen <koro.chen@mediatek.com>

BUG=chrome-os-partner:38746
TEST=build and boot oak-rev0

Change-Id: I1aa8d6571f8f6cbb1d4ed8bf89be4662c80bf01f