arm64, topology: Define JUNO energy and provide it to the scheduler
This is basically same hack we have in arm for TC2 were the energy
model tables are hardcoded in topology.c.

KP did almost the same. This is a slightly modified patch.

The fact that JUNO capacity scales nicely with frequency allows us
to correctly indexing the energy model tables with just to minor
modification to the measured values.

Signed-off-by: Juri Lelli <juri.lelli@arm.com>