EASv4 ARM64: arm64, topology: Define JUNO energy and provide it to the scheduler
This is basically same hack we have in arm for TC2 were the energy
model tables are hardcoded in topology.c.

KP did almost the same. This is a slightly modified patch.

The fact that JUNO capacity scales nicely with frequency allows us
to correctly indexing the energy model tables with just to minor
modification to the measured values.

Change-Id: I79f85715d0d1ca04295dbd3e7d735e46de484dca
Signed-off-by: Juri Lelli <juri.lelli@arm.com>